+++ App/Headers/All.h
@@ -1,0 +1,11 @@
+#ifndef ALLHEAD_INCLUDED
+#define ALLHEAD_INCLUDED
+
+//Callbacks para todos os Headers
+
+#include "File.h"
+#include "Data.h"
+#include "Menu.h"
+#include "Unissex.h"
+
+#endif
