+++ App/Headers/Data.h
@@ -1,0 +1,14 @@
+#ifndef DATA_INCLUDED
+#define DATA_INCLUDED
+
+
+typedef struct datanode{
+    int matricula;  //Armazenar a Matricula
+    char *nome; //Armazenar o Nome
+    char *sobrenome; //Armazenar o Sobrenome
+    char *email; //Armazenar o Email
+    char *telefone; //Armazena o Telefone
+    double salario; //Armazena o Salario
+}DataNode;
+
+#endif
